chore: Moved tooltips in separate file
