+++ src/ui.tsx
@@ -38,6 +38,7 @@ import { Labels } from "@/components/Labels";
 import { Download } from "@/download/UI";
 import { keydownListener } from "@/keybindings";
 import { Tools, Tool } from "@/tools";
+import { toolTips, minimapToolTips, annotationToolTips } from "./tooltips";
 
 const CONFIG = {
   PAN_AMOUNT: 20,
@@ -206,12 +207,6 @@ const styles = {
   svgSmall: { width: "18%", height: "auto" },
 };
 
-interface ToolTips {
-  name: string;
-  icon: string;
-  shortcut: string;
-}
-
 const HtmlTooltip = withStyles((t: Theme) => ({
   tooltip: {
     backgroundColor: "#FFFFFF",
@@ -231,100 +226,6 @@ class UserInterface extends Component<Props, State> {
   private imageFileInfo: ImageFileInfo | null;
 
   private canvasContainer: HTMLDivElement;
-
-  toolTips: ToolTips[] = [
-    {
-      name: "Select",
-      icon: require(`./assets/select-icon.svg`) as string,
-      shortcut: "A",
-    },
-    {
-      name: "Brush",
-      icon: require(`./assets/brush-icon.svg`) as string,
-      shortcut: "B",
-    },
-    {
-      name: "Eraser",
-      icon: require(`./assets/eraser-icon.svg`) as string,
-      shortcut: "E",
-    },
-    {
-      name: "Spline",
-      icon: require(`./assets/splines-icon.svg`) as string,
-      shortcut: "S",
-    },
-    {
-      name: "Magic Spline",
-      icon: require(`./assets/magic-spline-icon.svg`) as string,
-      shortcut: "M",
-    },
-
-    {
-      name: "Contrast",
-      icon: require(`./assets/contrast-icon.svg`) as string,
-      shortcut: `\\`,
-    },
-    {
-      name: "Brightness",
-      icon: require(`./assets/brightness-icon.svg`) as string,
-      shortcut: `/`,
-    },
-    {
-      name: "Channel",
-      icon: require(`./assets/channels-icon.svg`) as string,
-      shortcut: `C`,
-    },
-
-    {
-      name: "Annonation Label",
-      icon: require(`./assets/annotation-label-icon.svg`) as string,
-      shortcut: "L",
-    },
-  ];
-
-  minimapToolTips = [
-    {
-      name: "Minimise Map",
-      icon: require(`./assets/minimise-icon.svg`) as string,
-      shortcut: "ALT",
-      shortcutSymbol: "-",
-      styling: { marginRight: "120px", marginLeft: "15px" },
-    },
-
-    {
-      name: "Zoom In",
-      icon: require(`./assets/zoom-in-icon.svg`) as string,
-      shortcut: "CTRL",
-      shortcutSymbol: "+",
-      styling: { marginRight: "10px" },
-    },
-    {
-      name: "Zoom Out",
-      icon: require(`./assets/zoom-out-icon.svg`) as string,
-      shortcut: "CTRL",
-      shortcutSymbol: "-",
-      styling: { marginRight: "10px" },
-    },
-    {
-      name: "Fit to Page",
-      icon: require(`./assets/reset-zoom-and-pan-icon.svg`) as string,
-      shortcut: "CTRL",
-      shortcutSymbol: "]",
-    },
-  ];
-
-  annotationToolTips = [
-    {
-      name: "Add New Annotation",
-      icon: require(`./assets/new-annotation-icon.svg`) as string,
-      shortcutSymbol: "+",
-    },
-    {
-      name: "Clear Annotation",
-      icon: require(`./assets/delete-annotation-icon.svg`) as string,
-      shortcutSymbol: "-",
-    },
-  ];
 
   constructor(props: Props) {
     super(props);
@@ -736,7 +637,7 @@ class UserInterface extends Component<Props, State> {
       >
         <Grid container direction="row">
           <ButtonGroup size="small">
-            {this.annotationToolTips.map((toolTip) => (
+            {annotationToolTips.map((toolTip) => (
               <HtmlTooltip
                 key={toolTip.name}
                 title={
@@ -797,7 +698,7 @@ class UserInterface extends Component<Props, State> {
       >
         <Grid container direction="row">
           <ButtonGroup>
-            {this.toolTips.map((toolTip) => (
+            {toolTips.map((toolTip) => (
               <HtmlTooltip
                 key={toolTip.name}
                 title={
@@ -1078,7 +979,7 @@ class UserInterface extends Component<Props, State> {
               position: "relative",
             }}
           >
-            {this.minimapToolTips.map((minimapToolTip) => (
+            {minimapToolTips.map((minimapToolTip) => (
               <HtmlTooltip
                 key={minimapToolTip.name}
                 title={
